Linting & code formatting.

+++ flv_services/qdgw/src/utils.rs
@@ -24,7 +24,7 @@ impl Server {
         let mut symbol_db = self.symbol_manager.lock().await;
 
         // look up the table name
-        let res = symbol_db.get_symbol_table_name(exchange_id as u16, symbol_id);
+        let res = symbol_db.get_symbol_table_name(exchange_id, symbol_id);
 
         // Unlock the SymbolManager
         drop(symbol_db);
